refactor(template): move cn and formatPrice into lib/utils/ directory
Split lib/utils.ts into lib/utils/cn.ts and lib/utils/format.ts with
a barrel re-export in lib/utils/index.ts to preserve all existing
import paths and shadcn compatibility.
