+++ apps/template/lib/utils/index.ts
@@ -1,0 +1,2 @@
+export { cn } from "./cn";
+export { formatPrice } from "./format";
